backend/requirements.txt
# ===================================================================
# MomWatch AI - Backend Dependencies
# ===================================================================
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
aiohttp==3.9.1
httpx==0.26.0

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0

# Environment Management
python-decouple==3.8

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging
python-json-logger==2.0.7
